pointercaptureevents.html

<!DOCTYPE html>
<html>
<!--
	Test cases for Pointer Events v1 spec
-->
<head>
	<title>Pointer Events Capture Events Tests</title>
	<meta name="viewport" content="width=device-width">
	<link rel="author" href="mailto:rodrigos@outlook.com">
	<link rel="help" href="http://www.w3.org/wiki/PointerEvents/TestAssertions">
	<link rel="stylesheet" href="testharness.css"/>
	<!-- http://w3c-test.org/resources/testharness.js -->
	<script src="testharness.js"></script>
	<script src="pointer-prefix.js"></script>
	<script>
		var test_capture = async_test("Capture events");

		function run() {
			var target = document.getElementById("target");
			var capturedId = -1;
			var syncTestFlag;
			var pointerCaptured = false;
			var pointerReleased = false;

			function onPointerEvent(event) {
				if (pointerCaptured) {
					test(function() {
						assert_equals(event.type, pointerPrefix["gotpointercapture"],
							"First event to fire after setPointerCapture was " + pointerPrefix["gotpointercapture"])
					}, "First event to fire after setPointerCapture was " + pointerPrefix["gotpointercapture"]);
					pointerCaptured = false;
				}

				if (pointerReleased) {
					test(function() {
						assert_equals(event.type, pointerPrefix["lostpointercapture"],
								"First event to fire after releasePointerCapture was " + pointerPrefix["lostpointercapture"])
					}, "First event to fire after releasePointerCapture was " + pointerPrefix["lostpointercapture"]);
					pointerReleased = false;
				}
			}

			[
				"pointerdown",
				"pointerup",
				"pointercancel",
				"pointermove",
				"pointerover",
				"pointerout",
				"pointerenter",
				"gotpointercapture",
				"lostpointercapture"
			].forEach(function(eventName){
				on_event(target, pointerPrefix[eventName], onPointerEvent)
			});

			on_event(target, pointerPrefix["pointerdown"], function(event) {
				capturedId = event.pointerId;
				syncTestFlag = false;
				target[pointerPrefix["setPointerCapture"]](capturedId);
				pointerCaptured = true;
				syncTestFlag = true;
			});

			on_event(target, pointerPrefix["gotpointercapture"], function(event) {
				test(function(){
					assert_true(syncTestFlag, "gotpointercapture fired asynchronously");
				}, "gotpointercapture fired asynchronously");
				test(function() {
					assert_true(event.pointerId === capturedId,
						"gotpointercapture event received for pointerId " + capturedId);
				}, "gotpointercapture event received for pointerId " + capturedId);
			});

			on_event(target, pointerPrefix["pointerup"], function(event) {
				test(function() {
					assert_true(event.pointerId === capturedId,
						"pointerup event received for same id as pointer up " + capturedId);
				}, "pointerup event received for same id as pointer up " + capturedId);

				syncTestFlag = false;
				target[pointerPrefix["releasePointerCapture"]](capturedId);
				pointerReleased = true;
				syncTestFlag = true;
			});

			on_event(target, pointerPrefix["lostpointercapture"], function(event) {
				test(function(){
					assert_true(syncTestFlag, "lostpointercapture fired asynchronously");
				}, "lostpointercapture fired asynchronously");
				test(function() {
					assert_true(event.pointerId === capturedId,
						"lostpointercapture event received for pointerId " + capturedId);
				}, "lostpointercapture event received for pointerId " + capturedId);

				test_capture.done();
			});
		}

	</script>
	<style>
		#target {
			padding: 2em;
			background: yellow;
			border: 2px solid green;
			text-align: center;
			color: blue;
		}
	</style>
</head>
<body onload="run()">
	<h1>Pointer Events Capture Events Tests</h1>
	<div id="target">
		Click or tap here
l	</div>
	<div id="log"></div>
</body>
</html>
